requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
openpyxl>=3.1.0
PyYAML>=6.0
sentence-transformers>=2.2.0
openai>=1.0.0
gensim>=4.3.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.0.0